web/code/requirements.txt
Django==1.10.2
django-extensions==1.7.4
django-markdown==0.8.4
django-markup==1.1
django-mptt==0.8.6
django-mptt-admin==0.4.1
django-suit==0.2.21
docopt==0.6.2
docutils==0.12
jedi==0.9.0
Markdown==2.6.7
Pillow==3.4.2
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptpython==0.36
Pygments==2.1.3
python-creole==1.3.1
requests==2.11.1
six==1.10.0
smartypants==1.8.6
textile==2.3.4
wcwidth==0.1.7
Werkzeug==0.11.11
